+++ Arduino/robo2019/src/openmv.h
@@ -112,7 +112,7 @@ namespace openmv {
         }
     };
 
-    constexpr double pos2dir(const Position &pos) {
+    constexpr float pos2dir(const Position &pos) {
         return atan2(
             -float(pos.x) + 90, // = -(pos.x - 90)
             float(pos.y) - 70
